+++ src/components/grid/members/GridMembers.tsx
@@ -5,14 +5,14 @@ import StarIcon from '../../../components/icon/starIcon/starIcon';
 import TickIcon from '../../../components/icon/tickIcon/tickIcon';
 import TrashIcon from '../../../components/icon/trashIcon/trashIcon';
 import Alert from '../../alerts/standardAlert/StandardAlert';
-import TeamMember from '../../../utils/types/teamMember';
+import Investigator from '../../../utils/types/investigator';
 
 interface GridMembersProps {
   action?: boolean;
   actionClicked?: Function;
   height?: number;
   rowClick?: Function;
-  rows?: TeamMember[];
+  rows?: Investigator[];
 }
 
 export default function GridMembers({
@@ -100,7 +100,7 @@ export default function GridMembers({
           columns={getColumns()}
           height={height}
           onRowClick={rowClick}
-          testId="teamTableId"
+          testId="investigatorsTableId"
         />
       )}
       {!rows ||